pyproject.toml
[tool.poetry]
name = "08-qa-from-pdf"
version = "0.1.0"
description = ""
authors = ["Riyadh <riyadhgenai@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
jupyterlab = "^4.2.4"
langchain = "^0.2.11"
langchain-openai = "^0.1.20"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.10"
pypdf = "^4.3.1"
langchain-chroma = "^0.1.2"
langchain-groq = "0.1.5"
sentence-transformers = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
